restructured project, cleaned login
