+++ src/components/forms/LoginForm/initialValues.json
@@ -1,0 +1,4 @@
+{
+  "username": "",
+  "password": ""
+}
