+++ src/components/pages/LoginPage.jsx
@@ -1,0 +1,13 @@
+// Libraries
+import React from 'react'
+
+// Components
+import LoginForm from "./../forms/LoginForm/LoginForm"
+
+function LoginPage() {
+  <section>
+    <LoginForm />
+  </section>
+}
+
+export default LoginPage
