+++ src/config/index.js
@@ -1,5 +1,3 @@
-const config = {
+module.exports = {
   API_URL: 'https://anywhere-fitness-bwft5.herokuapp.com/api'
 }
-
-export default config
