--- src/styles/App.css
@@ -1,142 +1,0 @@
-*{
-    background-color:#232a47;
-    color: white;
-    font-family: 'Nunito', sans-serif;
-    font-family: 'Source Sans Pro', sans-serif;
-    padding:0%;
-    margin:0%;
-}
-.navbar-container{
-    display: flex;
-    justify-content: space-between;
-    margin:1%;
-    animation: 3s ease-in content;
-    
-}
-.menu{
-    width:10%;
-}
-.menu a{
-    padding-right: 8%;
-    text-decoration: none;
-}
-.menu a:hover{
-    color:rgb(2, 189, 117);
-}
-.homepage-container{
-    margin:10%;
-    width:100%;
-    padding:0% 2%;
-    display: flex;
-    flex-direction: column;
-    animation: 1s ease-out slideInLeft;
-}
-hr{
-    color:white;
-    width:50%;
-}
-.btn {
-    border: 1px solid white;
-    margin:4% 0%;
-    padding:1%;
-    width:20%;
-    font-size:1.2rem;
-    text-decoration: none;
-}
-.btn:hover{
-    color:rgb(2, 189, 117);
-    border: 1px solid rgb(2, 189, 117);
-}
-.content-container{
-    margin:2% 9%;
-    display: flex;
-    flex-direction: column;
-    animation: 1s ease-in content;
-}
-
-.landingpage-content{
-    display: flex;
-    justify-content: space-between;  
-   
-}
-.landingpage-content img{
-    width:40%;
-}
-.content{
-    width:60%;
-    margin:5%;
-    line-height: 2rem;;
-}
-.content button{
-    width: 20%;
-}
-.form {
-    display: flex;
-    flex-direction: column;
-    border-radius: 8px;
-    box-shadow: 1px 2px 8px rgba(0, 0, 0, 0.65);
-    height: 25%;
-    margin: 6rem auto 8.1rem auto;
-    width: 30%;
-    padding: 2% 3%;
-    animation: 0.5s ease-in content;
-}
-#signUp-form{
-    width: 39%;
-}
-
-.userCard {
-    display: flex;
-    flex-direction: column;
-    background: rgb(35,42,71);
-    border-radius: 8px;
-    box-shadow: 1px 2px 8px rgba(0, 0, 0, 0.65);
-    height: 250px;
-    margin: 6rem auto 8.1rem auto;
-    width: 300px;
-    padding: 12px 44px;
-    animation: 1s ease-in content;
-}
-.infoList {
-    font-size: 1.3em;
-    margin: 5%;
-}
-.input-container{
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-end;
-    font-size: 1.2rem;
-    line-height: 2rem;
-}
-
-input{
-    background-color: rgba(42,47,74,255);
-    border:rgba(42,47,74,255) ;
-    width: 95%;
-    height:2.1rem;
-}
-input:focus, textarea:focus{
-    background-color: rgb(19, 26, 49);
-    outline: none;
-    border: 1px solid rgb(2, 189, 117);
-}
-
-/* Animations */
-@keyframes slideInLeft {
-    0% {
-      opacity: 0;
-      transform: translateX(-5%)
-    }
-    100% {
-     opacity: 1;
-     transform: translateX(0)
-    }
-  }
-@keyframes content{
-    0%{
-        opacity: 0;
-    }
-    100%{
-        opacity: 1;
-    }
-}
